Controller has been created. + entity corrections

+++ src/main/java/com/avinty/hr/service/DepartmentService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface DepartmentService {
 
-    void deleteDepartment(Long id);
+    void deleteDepartment(int id);
 
 
     List<Department> searchDepartment(String name);
